Improvements for integers (#120)
Integer constants are now always parsed as BIGINT and integer types
(`SMALLINT`, `INTEGER` and `BIGINT`) are stored as an i64 in the Value
rather than storing them in the existing f64 which will lose precision
of the integer.

This allows use to raise a new SQLSTATE 22003 "numeric value out of
range" in most cases. However, this is not entirely complete as the
minimum value for BIGINT cannot be encoded corrected. See #116.

+++ generate-grammar.py
@@ -418,7 +418,7 @@ grammar_file.write("""fn parse_ast(node &EarleyNode) ?[]EarleyValue {
     if node.children.len == 0 {
         match node.value.name {
             '^integer' {
-                return [EarleyValue(node.value.end_column.value.int())]
+                return [EarleyValue(node.value.end_column.value)]
             }
             '^identifier' {
                 return [EarleyValue(new_identifier(node.value.end_column.value))]
